requirements.txt
click==6.6
ordereddict==1.1
py==1.4.31
pytest==2.9.2
python-dotenv==0.5.1
requests==2.11.1
